requirements.txt
typing-extensions
pycryptodome
coincurve
aiohttp[speedups]
jsonrpcclient
mnemonic
bip32
snapshottest
sphinx-rtd-theme
readthedocs-sphinx-ext
